usb-tds-sensor/requirements.txt
esptool
pyserial
mysql-connector-python 